+++ app/controllers/omniauth_callbacks_controller.rb
@@ -1,12 +1,6 @@
 class OmniauthCallbacksController < Devise::OmniauthCallbacksController
   include Devise::Controllers::Rememberable
 
-  PREVIOUSLY_BANNED_MESSAGE = "It appears that your previous %<community_name>s " \
-    "account was suspended. As such, we've taken measures to prevent you from " \
-    "creating a new account with %<community_name>s and its community. If you " \
-    "think that there has been a mistake, please email us at %<community_email>s, " \
-    "and we will take another look.".freeze
-
   # Rails actionpack only allows POST requests that come with an ORIGIN header
   # that matches `request.base_url`, it raises CSRF exception otherwise.
   # There is no way to allow specific ORIGIN values in order to securely bypass
@@ -101,11 +95,8 @@ class OmniauthCallbacksController < Devise::OmniauthCallbacksController
       flash[:alert] = user_errors
       redirect_to new_user_registration_url
     end
-  rescue Authentication::Authenticator::PreviouslyBanned => e
-    ForemStatsClient.increment("users.banned_signup_attempt", tags: ["error:#{e.class.name}"])
-    flash[:global_notice] = format(PREVIOUSLY_BANNED_MESSAGE,
-                                   community_name: SiteConfig.community_name,
-                                   community_email: SiteConfig.email_addresses[:contact])
+  rescue ::Authentication::Errors::PreviouslyBanned => e
+    flash[:global_notice] = e.message
     redirect_to root_path
   rescue StandardError => e
     Honeybadger.notify(e)
