--- app/services/authentication/errors.rb
@@ -1,12 +1,0 @@
-module Authentication
-  module Errors
-    class Error < StandardError
-    end
-
-    class ProviderNotFound < Error
-    end
-
-    class ProviderNotEnabled < Error
-    end
-  end
-end
